book-service/requirements.txt
django
djangorestframework
requests
PyMySQL
cryptography
